% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landfill.R
\docType{data}
\name{landfill}
\alias{landfill}
\title{Water Percolation Through A Landfill}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 974 rows and 48 columns.}
\usage{
landfill
}
\description{
A computer simulation of water moving through a landfill.  Detailed
explanation for the variables and the construction of the data set is found
in
\code{vignette("ensr-datasets", package = "ensr")}
}
\seealso{
\code{vignette("ensr-datasets", package = "ensr")}
}
\keyword{datasets}
