% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensr-pkg.R, R/ensr.R
\docType{package}
\name{ensr}
\alias{ensr}
\alias{ensr-package}
\title{Elastic Net SearcheR}
\usage{
ensr(x, y, alphas = seq(0.05, 0.95, by = 0.05), nfolds = 10L, foldid,
  ...)
}
\arguments{
\item{x}{\code{x} matrix as in \code{glmnet}.}

\item{y}{response \code{y} as in \code{glmnet}.}

\item{alphas}{a sequence alpha values}

\item{nfolds}{number of folds - default is 10. Although \code{nfolds}
    can be as large as the sample size (leave-one-out CV), it is not
    recommended for large datasets. Smallest value allowable is \code{nfolds=3}}

\item{foldid}{an optional vector of values between 1 and \code{nfold}
    identifying what fold each observation is in. If supplied,
    \code{nfold} can be missing.}

\item{...}{Other arguments that can be passed to \code{glmnet}}
}
\description{
Search for the values of alpha and lambda.
}
