% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda_max.R
\name{lambda_alpha_grid}
\alias{lambda_alpha_grid}
\title{Lambda Alpha Grid}
\usage{
lambda_alpha_grid(lambdas, alphas)
}
\arguments{
\item{lambdas}{a vector of max values for each alpha given}

\item{alphas}{a vector a alpha values corresponding to the max lambdas}
}
\description{
Construct a data frame with values for lambda and alpha with an indicator to
know if the model is worth fitting.
}
\examples{

data(tbi)
Xmat <- model.matrix( ~ . - injury1 - injury2 - injury3 - 1, data = tbi)
Yvec <- matrix(tbi$injury1, ncol = 1)
alphas <- seq(0, 1, length = 20)
test <- lambda_alpha_grid(alphas = alphas, lambdas = lambda_max(Yvec, Xmat, alpha = alphas))

ggplot2::ggplot() +
  ggplot2::theme_bw() +
  ggplot2::aes_string(x = "a", y = "log10(l)") +
  ggplot2::geom_path(data = test$top) +
  ggplot2::geom_point(data = test$lgrid)

}
