% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensr.R
\name{predict}
\alias{predict}
\alias{predict.ensr}
\alias{coef.ensr}
\title{Predict Methods for ensr objects}
\usage{
\method{predict}{ensr}(object, ...)

\method{coef}{ensr}(object, ...)
}
\arguments{
\item{object}{a \code{ensr} object}

\item{...}{other arguments passed along to \code{predict}}
}
\description{
Using either the \code{lambda.min} or \code{lambda.1se}, find the preferable
model from the \code{ensr} object and return a prediction.
}
\details{
The \code{glmnet::predict} argument \code{s} is ignored if specified and
attempted to be passed via \code{...}.  The value of \code{s} that is passed
to \code{glmnet::predict} is determined by the value of \code{lambda.min} or
\code{lambda.1se} found from a call to \code{\link{preferable}}.
}
\seealso{
\code{\link[glmnet]{predict}}
}
