% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensr.R
\name{preferable}
\alias{preferable}
\title{Preferable Elastic Net Model}
\usage{
preferable(object, cve = "min", ...)
}
\arguments{
\item{object}{an ensr object}

\item{cve}{a character string, by default using \code{"min"} will use the
\code{lambda.min} for selecting a preferable model.  \code{"1se"} will use
\code{lambda.1se}.}

\item{...}{not currently used.}
}
\description{
Find the preferable Elastic Net Model from an ensr object.  The preferable
model is defined as the model with the lowest mean cross validation error and
largest alpha value.
}
