% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensr.R
\name{preferable}
\alias{preferable}
\title{Preferable Elastic Net Model}
\usage{
preferable(object, ...)
}
\arguments{
\item{object}{an ensr object}

\item{...}{not currently used.}
}
\description{
Find the preferable Elastic Net Model from an ensr object.  The preferable
model is defined as the model with the lowest mean cross validation error and
largest alpha value.
}
